+++ Sources/KeyboardKit/_Deprecated/KeyboardLocale+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension KeyboardLocale {
+    
+    @available(*, deprecated, renamed: "identifier")
+    var key: String { id }
+}